refactor: add database adapter to switch between data sources
